replace Class with utils
