+++ src/lib/userStorage/UserProfileStorage.js
@@ -11,8 +11,8 @@ import isEmail from '../validators/isEmail'
 import isMobilePhone from '../validators/isMobilePhone'
 import type { UserModel } from './UserModel'
 import { getUserModel } from './UserModel'
-import { UserProfileStaticMethods } from './UserProfileStaticMethods'
 import type { FieldPrivacy, ProfileField } from './UserStorageClass'
+import { isValidValue, maskField } from './utlis'
 
 const logger = pino.child({ from: 'UserProfileStorage' })
 
@@ -287,7 +287,7 @@ export class UserProfileStorage implements ProfileStorage {
         display = '******'
         break
       case 'masked':
-        display = UserProfileStaticMethods.maskField(field, value)
+        display = maskField(field, value)
 
         //undo invalid masked field
         if (display === value) {
@@ -473,7 +473,7 @@ export class UserProfileStorage implements ProfileStorage {
     const validatedFields = await Promise.all(
       fields.map(async field => ({
         field,
-        valid: await UserProfileStaticMethods.isValidValue(field, profile[field], true),
+        valid: await isValidValue(field, profile[field], true),
       })),
     )
